\name{Poptri.genind}
\alias{Poptri.genind}
\alias{Poptri.baseline.env}
\alias{Poptri.future.env}
\alias{Poptri.loc}
\alias{Poptri.freq.baseline}
\alias{Poptri.freq.future}
\alias{Poptri.1985to2085}



\docType{data}

\title{Example data sets for Populus trichocarpa.}

\description{
These data sets include the same data that were used by Blumstein et al. 2020 to document their protocol. Genomic data has been processed further into the \code{\link[adegenet]{genind}} format. Data sets Poptri.freq.baseline, Poptri.freq.future and Poptri.1985to2085 were obtained with the calibration and prediction methods of this package. They are included to make the documentation of the various plotting methods shorter, but also show the types of results someone can obtain.
}

\usage{
data(Poptri.genind)
data(Poptri.baseline.env)
data(Poptri.future.env)
data(Poptri.loc)
data(Poptri.freq.baseline)
data(Poptri.freq.future)
data(Poptri.1985to2085)
}


\references{
Blumstein et al. 2020. Protocol for Projecting Allele Frequency Change under Future Climate Change at Adaptive-Associated Loci. \doi{10.1016/j.xpro.2020.100061}
}

\examples{
data(Poptri.genind)
}
\keyword{datasets}
